Read all LLDP neighbours even if one fails
Currently in the readLLDP function, when looping over all neighbours in nbrList, if createR2RLink for one neighbour fails, the whole function returns which means that the rest of the neighbors in nbrList are never processed. I think it makes more sense to continue the loop and create R2R links for other neighbors. This patch proposes to do that while still returning false if one createR2RLink fails.

Change-Id: I18aeb55df6c445479e4a9b80e8d90848025480bb